\name{DataSelection}
\alias{DataSelection}
\title{Selection in Data}
\usage{
  DataSelection(data, column = 1L)
}
\arguments{
  \item{data}{\code{\link{mutaframe}} of the
  dataset/pipeline}

  \item{column}{Column index of selection variable in data}
}
\value{
  An \code{\link{ItemSelection}} reflecting the selection
  in the data
}
\description{
  Implement a selection model against a dataset/pipeline
}
\author{
  Michael Lawrence
}

