\name{is.mutaframe}
\alias{is.mutaframe}
\title{Test for mutaframes}
\usage{
  is.mutaframe(x)
}
\arguments{
  \item{x}{an object to check}
}
\value{
  \code{TRUE} if \code{x} is an instance of a class that
  inherits from \code{mutaframe}; otherwise, \code{FALSE}
}
\description{
  Tests whether an object is a \code{mutaframe}
}

