\name{mutaframe}
\alias{mutaframe}
\title{Create a mutaframe, a mutable data.frame}
\usage{
  mutaframe(..., row.names = NULL)
}
\arguments{
  \item{...}{Objects to coerce to a mutaframe and combine
  column-wise}

  \item{row.names}{optional, the character vector of row
  names}
}
\value{
  a mutaframe
}
\description{
  Create a mutaframe, a mutable data.frame
}

