\name{raw_binding}
\alias{raw_binding}
\title{Generate binding for raw values}
\usage{
  raw_binding(mf, name, data)
}
\arguments{
  \item{mf}{mutaframe}

  \item{name}{name}

  \item{data}{vector to store}
}
\value{
  named list of binding functions
}
\description{
  Generate binding for raw values
}

