\name{remove_listener}
\alias{remove_listener}
\title{Remove a listener, identified by the ID returned by
\code{\link{add_listener}}.}
\usage{
  remove_listener(mf, id)
}
\arguments{
  \item{mf}{mutaframe}

  \item{id}{value returned by \code{add_listener} when
  originally connecting the handler}
}
\description{
  Remove a listener, identified by the ID returned by
  \code{\link{add_listener}}.
}

